Restrinjo comandos a un canal

+++ comandos/first/nowater.js
@@ -13,19 +13,20 @@ module.exports = class WaterNoReminder extends Command {
     }
 
     async run(message) {
-
-        Reminder.findOneAndDelete({
-            username: message.author.username,
-            userID: message.author.id
-        }, (err, user) => {
-            if (err) console.log(err);
-            if (!user) {
-                message.reply('No estas en la base de datos. Para entrar, !water');
-            } else{
-             message.reply('Removido de la base de datos!');
-             console.log(user);
-            }
-        });
-
+        if (message.channel.id != '596082817734148157') message.delete();
+        else{
+            Reminder.findOneAndDelete({
+                username: message.author.username,
+                userID: message.author.id
+            }, (err, user) => {
+                if (err) console.log(err);
+                if (!user) {
+                    message.reply('No estas en la base de datos. Para entrar, !water');
+                } else{
+                message.reply('Removido de la base de datos!');
+                console.log(user);
+                }
+            });
     }
+}
 };
